Completed lab 3
This lab was a big pain, because colab continuously kicked me off. I
needed to update it to save off results periodically. I also tried
running it locally to get a continuous run, and thus had to add new
dependencies, but even that did not run to completion.

I finally got things to work, although the results won't be the same
as if I had run it with a single random seed for one run.